requirements.txt
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
langchain-chroma>=0.1.0,<1.0
langchain-huggingface>=0.0.1,<1.0
langchain-community>=0.0.30,<1.0
langchain-text-splitters>=0.0.1,<1.0
